Restore old versions of EVP_{CIPHER,MD}_CTX_dup symbols to avoid ABI breakage
This change is based on a patch by Sahana Prasad from the OpenSSL PR,
addressing issues mentioned in the PR comments.  Also, there is
important difference from the PR, the default versioning of these
symbols is OPENSSL_3.2.0, aligning with the currently released OpenSSL
3.2.1.  This is the safest approach to the issue for distributions.

Although GCC 10 supports the symver attribute, it does not support
multiple versions for one symbol, a feature available since GCC 11.

Fixes: https://github.com/openssl/openssl/issues/23534
Link: https://github.com/openssl/openssl/pull/23554

Signed-off-by: Gabi Falk <gabifalk@gmx.com>